use a consistent version of bikeshed
bikeshed changes frequently. we might all be running different versions
of it; this means that rebuilding index.html on different machines might
result in different versions of the site.

For consistency, we should lock requirements.txt to a particular
version.  in this commit I've picked the sha1 for latest master at the
current time.