use `:` and base64url, not `;` and base64
We're taking inspiration from [RFC 6920][], which in section 5 suggests
that URL path parts can take the form of their alg-val construction.  An
example of this is: sha-256;f4OxZX_x_FO5LcGBSKHWXfwtSx-j1ncoSt3SABJtkGk
-- it uses a hash algorithm name from [a registry][rfc6920-hash-reg], it
uses the semicolon character as a separator, and it uses base64url with
no padding characters.

However, semicolons have a number of issues: they're a reserved
character in a URL (sometimes used for so-called matrix parameters), and
they are special characters in bash so can mess up curl calls if you
don't quote the URL.

Base64url, which uses the url-safe characters "-_" instead of "+/",
makes sense, and it also makes sense to omit the padding character as
the hash is a known size (and the "=" padding character is messy in URLs
as it has a special meaning in query strings and matrix parameter
strings).

[RFC 6920]: https://tools.ietf.org/html/rfc6920
[rfc6920-hash-reg]: https://www.iana.org/assignments/named-information/named-information.xhtml