rename "tree-size" to "total-entries"
"tree-size" and "total-entries" both mean exactly the same thing.  We
shouldn't have two different words for one thing.  This commit renames
all instances of "tree-size" to be "total-entries" instead.

This is technically a breaking change to the Register Proof Resource,
but as we haven't implemented this part of it, we won't actually break
anyone.